backend/requirements.txt
Flask>=2.1
Pillow>=10.0
